+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## 2.3.0 - 2023-01-23
 
-- Added `shell.shellHook` and `pkgs` options in offchain module
+- Added `shell.shellHook`, `pkgs`, and `runtime.exposeConfig` options in offchain module
 
 ## 2.2.2 - 2023-01-16